requirements.txt
pandas
openpyxl
streamlit
plotly-express
weasyprint